Add option to use internal storage for cache
Some users have reported an issue with the cache resulting in an error message "cache is not available" in the gallery. However, I have not been able to replicate this issue on my own devices. It seems that on some devices, external files directory or external cache directory is not automatically created, which is causing the problem. To address this, I have added an option to use internal storage for the cache instead, as this storage is always available.